+++ paynow/__version__.py
@@ -1,0 +1,9 @@
+# 88""Yb    db     dP""b8 88  dP    db     dP""b8 888888
+# 88__dP   dPYb   dP   `" 88odP    dPYb   dP   `" 88__
+# 88"""   dP__Yb  Yb      88"Yb   dP__Yb  Yb  "88 88""
+# 88     dP""""Yb  YboodP 88  Yb dP""""Yb  YboodP 888888
+
+
+VERSION = (1, 0, 2)
+
+__version__ = '.'.join(map(str, VERSION))
