--- paynow/test.py
@@ -1,3 +1,0 @@
-items = [['a',1.03],['b',2],['b',2],['b',2],['b',2]]
-
-print(float(sum([items[index][1] for index in range(len(items))])))
