+++ setup.py
@@ -1,22 +1,119 @@
-from setuptools import setup, find_packages
+# #!/usr/bin/env python
+# -*- coding: utf-8 -*-
 
-from codecs import open
-from os import path
+# Note: To use the 'upload' functionality of this file, you must:
+#   $ pip install twine
 
-here = path.abspath(path.dirname(__file__))
+import io
+import os
+import sys
+from shutil import rmtree
 
-with open(path.join(here, 'README.md'), encoding='utf-8') as f:
-    long_description = f.read()
+from setuptools import find_packages, setup, Command
 
+# Package meta-data.
+NAME = 'paynow'
+DESCRIPTION = 'Paynow Python SDK'
+URL = 'https://gitlab.com/paynow-developer-hub/Paynow-Python-SDK'
+EMAIL = 'pkg-dev@webdevworld.com'
+AUTHOR = 'WebDev Projects'
+REQUIRES_PYTHON = '>=3.6.0'
+VERSION = None
+
+
+# What packages are required for this module to be executed?
+REQUIRED = [
+    'requests', 
+]
+
+# What packages are optional?
+EXTRAS = {
+    # 'fancy feature': ['django'],
+}
+
+# The rest you shouldn't have to touch too much :)
+# ------------------------------------------------
+# Except, perhaps the License and Trove Classifiers!
+# If you do change the License, remember to change the Trove Classifier for that!
+
+here = os.path.abspath(os.path.dirname(__file__))
+
+# Import the README and use it as the long-description.
+# Note: this will only work if 'README.md' is present in your MANIFEST.in file!
+try:
+    with io.open(os.path.join(here, 'README.md'), encoding='utf-8') as f:
+        long_description = '\n' + f.read()
+except FileNotFoundError:
+    long_description = DESCRIPTION
+
+# Load the package's __version__.py module as a dictionary.
+about = {}
+if not VERSION:
+    with open(os.path.join(here, NAME, '__version__.py')) as f:
+        exec(f.read(), about)
+else:
+    about['__version__'] = VERSION
+
+
+class UploadCommand(Command):
+    """Support setup.py upload."""
+
+    description = 'Build and publish the package.'
+    user_options = []
+
+    @staticmethod
+    def status(s):
+        """Prints things in bold."""
+        print('\033[1m{0}\033[0m'.format(s))
+
+    def initialize_options(self):
+        pass
+
+    def finalize_options(self):
+        pass
+
+    def run(self):
+        try:
+            self.status('Removing previous builds…')
+            rmtree(os.path.join(here, 'dist'))
+        except OSError:
+            pass
+
+        self.status('Building Source and Wheel (universal) distribution…')
+        os.system('{0} setup.py sdist bdist_wheel --universal'.format(sys.executable))
+
+        self.status('Uploading the package to PyPI via Twine…')
+        os.system('twine upload dist/*')
+
+        self.status('Pushing git tags…')
+        os.system('git tag v{0}'.format(about['__version__']))
+        os.system('git push --tags')
+        
+        sys.exit()
+
+
+# Where the magic happens:
 setup(
-    name='paynow',  # Required
-    version='1.0.2',  # Required
-    description='Paynow Python SDK',  # Required
-    long_description=long_description,  # Optional
-    long_description_content_type='text/markdown',  # Optional (see note above)
-    url='https://gitlab.com/paynow-developer-hub/Paynow-Python-SDK',  # Optional
-    author='WebDev Projects',  # Optional
-    author_email='pkg-dev@webdevworld.com',  # Optional
+    name=NAME,
+    version=about['__version__'],
+    description=DESCRIPTION,
+    long_description=long_description,
+    long_description_content_type='text/markdown',
+    author=AUTHOR,
+    author_email=EMAIL,
+    python_requires=REQUIRES_PYTHON,
+    url=URL,
+    packages=find_packages(exclude=['contrib', 'docs', 'tests']),  # Required
+    # If your package is a single module, use this instead of 'packages':
+    # py_modules=['mypackage'],
+
+    # entry_points={
+    #     'console_scripts': ['mycli=mymodule:cli'],
+    # },
+    install_requires=REQUIRED,
+    extras_require=EXTRAS,
+    include_package_data=True,
+    license='MIT',
     classifiers=[  # Optional
         'Development Status :: 5 - Production/Stable',
 
@@ -35,16 +132,17 @@ setup(
         'Programming Language :: Python :: 3.6',
     ],
     keywords='paynow-api paynow-zimbabwe paynow-zimbabwe-api paynow-zimbabwe-sdk',  # Optional
-    packages=find_packages(exclude=['contrib', 'docs', 'tests']),  # Required
-    # This field lists other packages that your project depends on to run.
-    # Any package you put here will be installed by pip when your project is
-    # installed, so they must be valid existing projects.
-    #
-    # For an analysis of "install_requires" vs pip's requirements files see:
-    # https://packaging.python.org/en/latest/requirements.html
-    install_requires=['requests'],  # Optional
     project_urls={  # Optional
         'Bug Reports': 'https://gitlab.com/paynow-developer-hub/Paynow-Python-SDK/issues',
         'Source': 'https://gitlab.com/paynow-developer-hub/Paynow-Python-SDK',
     },
+    # $ setup.py publish support.
+    cmdclass={
+        'upload': UploadCommand,
+    },
 )
+
+
+
+
+
